+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-## [0.1.1] - 2025-04-25
+## [0.1.1] - 2025-05-02
 
+- add SpanContext class to facilitate executing without request context
+- expose function to get tracing context trace flags
 - fix readme
 
 ## [0.1.0] - 2025-04-24